Update application configuration and improve user messages
This commit makes considerable changes to the application configuration and the user interface. The docker-compose file version is updated, and the application now runs on port 5001. Flask CORS is added to the main application to handle cross-origin requests. User-facing messages have also been updated to be more user-friendly and informative. Several environment configurations have been defined in the Dockerfile and a .env file has been created. Additionally, unnecessary components on the base template have been removed.